refactor plugin unit tests
